*** empty log message ***
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@1209 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/quakedef.h
@@ -230,7 +230,7 @@ void Host_Shutdown(void);
 void VARGS Host_Error (char *error, ...);
 void VARGS Host_EndGame (char *message, ...);
 qboolean Host_SimulationTime(float time);
-void Host_Frame (float time);
+void Host_Frame (double time);
 void Host_Quit_f (void);
 void VARGS Host_ClientCommands (char *fmt, ...);
 void Host_ShutdownServer (qboolean crash);
